# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel File API",
        "description": "API documentation for File Management",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/files/import": {
            "post": {
                "tags": [
                    "Files"
                ],
                "summary": "Import data from Excel file",
                "operationId": "92e59e6886bbf3d059710bb421c0c882",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "description": "Excel file to import",
                                        "type": "file"
                                    },
                                    "table": {
                                        "description": "Table name to import data",
                                        "type": "string",
                                        "enum": [
                                            "products",
                                            "users"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Import successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Import dữ liệu thành công"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lỗi validation dữ liệu"
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Có lỗi xảy ra khi import"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/files/export": {
            "post": {
                "tags": [
                    "Files"
                ],
                "summary": "Export data to Excel file",
                "operationId": "954248cbd5332198f803d330fb0c7191",
                "parameters": [
                    {
                        "name": "table",
                        "in": "query",
                        "description": "Table name to export data",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "products",
                                "users"
                            ]
                        }
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "properties": {
                                "is_active": {
                                    "type": "boolean"
                                },
                                "status": {
                                    "type": "integer"
                                },
                                "start_date": {
                                    "type": "string",
                                    "format": "date"
                                },
                                "end_date": {
                                    "type": "string",
                                    "format": "date"
                                },
                                "price_from": {
                                    "type": "number"
                                },
                                "price_to": {
                                    "type": "number"
                                },
                                "quantity_from": {
                                    "type": "integer"
                                },
                                "quantity_to": {
                                    "type": "integer"
                                },
                                "sort_field": {
                                    "type": "string"
                                },
                                "sort_direction": {
                                    "type": "string",
                                    "enum": [
                                        "asc",
                                        "desc"
                                    ]
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Export successful",
                        "content": {
                            "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid table",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Bảng không hợp lệ"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Export thất bại"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}